requirements.txt
Django==1.8
Pillow==2.8.1
dj-database-url==0.3.0
dj-static==0.0.6
django-sirtrevor
django-toolbelt==0.0.1
easy-thumbnails==2.2
gunicorn==19.3.0
html2text
psycopg2==2.6
pytils
static3==0.5.1
django-storages
boto
